Allow passing `bytes` type into `write_bytearray`
While it may be slightly odd to use `bytes` instead of `bytearray` to be
passed into a function for writing bytearrays, however it actually makes
a lot of sense.

Type-wise, `bytearray` is actually a subtype of bytes, so bytearrays can
still be passed with the annotation being `bytes` type.

The reason using `bytes` makes more sense is because this type also
supports `__len__`, and can be sent over as an array of individual
bytes. There is therefore no reason to disallow this type, and it does
allow us to avoid needlessly casting/converting bytes to bytearrays.